libs/python/mythica-api/pyproject.toml
[tool.poetry]
name = "mythica-api"
version = "0.1.0"
description = "A client library for accessing the Mythica API"
authors = ["Jacob Repp <jacobrepp@gmail.com>"]
readme = "README.md"
packages = [
    {include = "mythica_api"},
]
include = ["README.md", "CHANGELOG.md", "mythica_api/py.typed"]


[tool.poetry.dependencies]
python = "^3.8"
httpx = ">=0.20.0,<0.28.0"
attrs = ">=21.3.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
